created types directory
