+++ src/types/activity.ts
@@ -1,0 +1,9 @@
+export interface Activity {
+    id: number
+    type: string
+    title: string
+    status: string
+    statusType: string
+    date: string
+    action: string
+}
