+++ src/types/index.ts
@@ -1,0 +1,11 @@
+// User and Authentication types
+export * from './user'
+
+// Assessment and Test types
+export * from './assessment'
+
+// Category and Problem types
+export * from './category'
+
+// Activity and Resource types
+export * from './activity'
